developer utility | introduce util/develop/rule-compare.php
this tool is checking per Rule if SOURCE / DESTINATION / SERVICE value resolution is identical for two configuration files;
file1= -> always the original Filename
file2= -> manipulate/optimised configuration Filename; e.g. after type=address-merger
